+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 ### Other changes
 
   * added `stubbles\lang\Result`
+  * `@ImplementedBy` can now be speficied per runtime mode
   * it is not required any more to mark constructors with `@Inject` when they should be subject to dependency injection
   * added `stubbles\peer\IpAddress::createSocket()` and `stubbles\peer\IpAddress::createSecureSocket()`
   * improved speed for `stubbles\peer\http\HttpUri::hasDnsRecord()` by checking only relevant DNS records
@@ -61,7 +62,6 @@
 ------------------
 
   * fixed annotation string values which contained both ' and "
-
 
 
 5.5.0 (2015-05-06)